+++ src/FolkerKinzel.VCards/md/ReleaseNotes.md
@@ -1,5 +1,4 @@
-- Fixes an issue that a "geo" URI was not parsed if this "geo" URI was masked.
-- Dependency update
+- Dependency update: Uses `FolkerKinzel.DataUrls 2.0.0`
 
 &nbsp;
 >**Project reference:** On some systems, the content of the CHM file in the Assets is blocked. Before opening the file right click on the file icon, select Properties, and **check the "Allow" checkbox** - if it is present - in the lower right corner of the General tab in the Properties dialog.